Code pimping for Solaris
It seems as if Solaris was a bit pickier about type casting in its libc than FreeBSD. But no problem, in return, whoop (or what currently exists of it) builds now without warnings on Solaris 10/SPARC (sunos 5.10).

+++ src/lib/daemon/daemon.c
@@ -117,7 +117,7 @@ daemon_init(const char *program, int facility, const char *pidfile)
 		syslog(LOG_ERR, "Fatal error in fopen(): %s\n", strerror(errno));
 		exit(EXIT_FAILURE);
 	}
-	fprintf(pid_fp, "%d", pid);
+	fprintf(pid_fp, "%d", (int)pid);
 	fclose(pid_fp);
 }
